LPS-159413 Move to utils
